ADD: Added a new file just for functional testing on a scratch pad when working with interview cake problems. MOD: Changed some types in hacker rank loops I was using ADD: Changes to project file to add the inclusion of the new ScratchPad header MOD: Code in main function for testing new scratch pad functions

+++ TestConsole/HackerRank.h
@@ -31,12 +31,12 @@ int minimumSwaps(std::vector<int>& arr)
 	int minimumSwapsMade = 0;
 	std::cout << "traversing the arr" << std::endl;
 
-		for (int i = 0; i < arr.size(); i++)
+		for (size_t i = 0; i < arr.size(); i++)
 		{
 			// complete sorting operation
 			if (arr[i] != (i + 1))
 			{
-				for (int j = 0; j < arr.size(); j++)
+				for (size_t j = 0; j < arr.size(); j++)
 				{
 					if (arr[j] == (i + 1))
 					{
